requirements.txt
pandas
grpcio
grpcio-tools
jupyter
git+https://github.com/bsc-dom/dataclay
